+++ src/App.tsx
@@ -4,8 +4,8 @@ import { pokemonMessage } from "./messages";
 import loading from "./media/loading.svg";
 import { OpponentsTeamDisplay } from "./components/OpponentsTeamDisplay";
 import { getBattleType, isRandomBattle, isDevelopmentMode } from "./functions";
-import { NotPokemonShowdownErrorScreen } from "./components/NotPokemonShowdownErrorScreen";
-import { NotInBattleErrorScreen } from "./components/NotInBattleErrorScreen";
+import { NotPokemonShowdownErrorScreen } from "./components/ErrorScreens/NotPokemonShowdownErrorScreen";
+import { NotInBattleErrorScreen } from "./components/ErrorScreens/NotInBattleErrorScreen";
 import { TypeWriterContainer } from "./styles/TypeWriterContainer";
 import { AppDisplay, Button, Refresh, RefreshButton } from "./styles/AppStyles";
 import testDS from "./functions/testObjects";
