+++ src/components/ErrorScreens/NotInBattleErrorScreen.tsx
